Move util stuff into util dir
