+++ extension/src/util/input.ts
